Task #7

+++ 0x17-doubly_linked_lists/1-dlistint_len.c
@@ -6,7 +6,7 @@
  */
 size_t dlistint_len(const dlistint_t *h)
 {
-	int x = 0;
+	size_t x = 0;
 
 	while (h)
 	{
